backend/pyproject.toml
[tool.poetry]
name = "app-py3-12-fastapi"
version = "0.1.0"
description = "Template backend project"
authors = ["like <1971464771@qq.com>"]
readme = "README.md"
packages = [{include = "app_py3"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-dotenv = "1.0.1"
prometheus-fastapi-instrumentator = "7.0.0"
markdown = "^3.6"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
